+++ CHANGELOG.md
@@ -3,8 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Self test for case setup failure.
 ### Fixed
-- Repeat count in example in Readme
+- Repeat count in example in Readme.
 
 ## [1.7.4] - 2015-12-17
 ### Added